refactor(sprite): reorganize sprite implementation code structure
- Extract bubble/communication methods to sprite_bubble.go
  * Move Ask, Say, Think, Quote methods from sprite_impl.go

- Extract render-related methods to sprite_render.go
  * Move visibility control (Hide/Show/Visible)
  * Move costume methods (SetCostume, CostumeName, CostumeIndex)
  * Move graphic effects (SetGraphicEffect, ChangeGraphicEffect, ClearGraphicEffects)
  * Move layer control (SetLayer)

- Integrate animation event methods into animation component
  * Add registerOnAnimationLooped/Finished wrappers in component_animation.go
  * Update gdspx.go to use animation component methods
  * Update sprite_impl.go Destroy() to use animation component

- Improve naming conventions
  * Rename _onTouchStart to addTouchStartHandler for better clarity
